Added a Sensors Class
Several changes have been made:

1.   Added the Sensors class which will make it easier to use on the
main file. It also setups for possibilities of adding more or different
sensors for the future.

2.     There's also a spreadsheet with the anallysis of the last
readgins in each trial we did in the Growth Lab.

3.      The DHT files are no longer in the main directory. You need to
add the library to your arduino libraries folder. The library is in the
TestFiles/TempHumid/ folder. Just add that zip file to your libraries
and you shall get no errors.