Added Reading Checks for Warmup State
A new function checks the last 5 readings. This determines whether the
readings obtained are accurate or if the sensors are still stabilizing.
Do note this will get the Warm up screen stuck if the sensors are not
plugged in. You must include the Statistics library included in the Test
files and Libraries directory